Verificar archivo..

+++ proyecto/perfil/views.py
@@ -80,34 +80,38 @@ class UploadView(CreateView):
     model = Upload
     form_class = UploadForm
     def form_valid(self, form):
-        form.save()
+        #form.save()
         nombre = self.request.user.username
         currentpassword = self.request.user.password
         archivo = form.cleaned_data.get('upload_file').read()
         archivo_nombre = form.cleaned_data.get('upload_file').name
         password = form.cleaned_data.get('password')
         matchcheck= check_password(password, currentpassword)
-        print(matchcheck)
+
         if matchcheck is not False:
             llave_aes = generar_llave_aes_from_password(password)
             iv = b"M\xb0%\xafd)\xe7\x11@7'\xb0\xcc\xc9\x81\xe2"
             path_privada_des ='./llaves/' + nombre + 'privada.pem'
             path_privada_cif = './llaves/' +nombre + 'privada.pem.cif'
-            with open(path_privada_des, 'wb') as salida_des:
-                contenido = descifrar(regresar_b_arch(path_privada_cif), llave_aes, iv)
-                salida_des.write(contenido)
-            llave_priv = convertir_bytes_llave_privada(contenido)
-            firma = firmar(llave_priv, archivo)
-            nombre_firma = './firmas/firma' + nombre + archivo_nombre
-            remove(path_privada_des)
-            with open(nombre_firma, 'wb') as firmado:
-                contenido = firma
-                firmado.write(contenido)
-            response = HttpResponse(open(nombre_firma, 'rb').read())
-            response['Content-Type'] = 'text/plain'
-            response['Content-Disposition'] = "attachment; filename=firma"+nombre +archivo_nombre
-            remove(nombre_firma)
-            return response
+            if os.path.isfile(path_privada_cif):
+                with open(path_privada_des, 'wb') as salida_des:
+                    contenido = descifrar(regresar_b_arch(path_privada_cif), llave_aes, iv)
+                    salida_des.write(contenido)
+                llave_priv = convertir_bytes_llave_privada(contenido)
+                firma = firmar(llave_priv, archivo)
+                nombre_firma = './firmas/firma' + nombre + archivo_nombre
+                remove(path_privada_des)
+                with open(nombre_firma, 'wb') as firmado:
+                    contenido = firma
+                    firmado.write(contenido)
+                response = HttpResponse(open(nombre_firma, 'rb').read())
+                response['Content-Type'] = 'text/plain'
+                response['Content-Disposition'] = "attachment; filename=firma"+nombre +archivo_nombre
+                remove(nombre_firma)
+                return response
+            else:
+                messages.error(self.request, 'Llaves expiradas genera nuevas')
+                return redirect('/upload')
         else:
             messages.error(self.request,'La contraseña que ingresaste no coincide, intenta de nuevo')
             return redirect('/upload')
@@ -117,25 +121,27 @@ class VerifySign(CreateView):
     form_class = VerifySignForm
     def form_valid(self, form):
         users = [str(user) for user in User.objects.all()]
-        form.save()
+        #form.save()
 
         usuario = form.cleaned_data.get('usuario')
         if usuario in users:
             path_publica = './llaves/' + usuario + 'publica.pem'
-            llave_publica = convertir_bytes_llave_publica(regresar_b_arch(path_publica))
-            archivo = form.cleaned_data.get('upload_file').read()
-            firma = form.cleaned_data.get('upload_firma').read()
-            try:
-                llave_publica.verify(firma,archivo,ec.ECDSA(hashes.SHA256()))
-                print('La firma es válida')
-                messages.success(self.request, 'La firma es válida.')
-                return redirect('/verificar-firma')
-            except InvalidSignature:
-                print("La firma es inválida")
-                messages.error(self.request, 'La firma es inválida.')
-                return redirect('/verificar-firma')
+            if os.path.isfile(path_publica):
+                llave_publica = convertir_bytes_llave_publica(regresar_b_arch(path_publica))
+                archivo = form.cleaned_data.get('upload_file').read()
+                firma = form.cleaned_data.get('upload_firma').read()
+                try:
+                    llave_publica.verify(firma,archivo,ec.ECDSA(hashes.SHA256()))
+                    print('La firma es válida')
+                    messages.success(self.request, 'La firma es válida.')
+                    return redirect('/verificar-firma')
+                except InvalidSignature:
+                    print("La firma es inválida")
+                    messages.error(self.request, 'La firma es inválida.')
+                    return redirect('/verificar-firma')
+            else:
+                messages.error(self.request, 'Por favor genera nueva llaves')
         else:
-
             messages.error(self.request, 'el usuario no existe, no tiene firmas')
             return redirect('/verificar-firma')
